--- backend/src/api/resources/auth/userRoute.js
@@ -1,14 +1,0 @@
-// /routes/api/user.route.js
-var express = require('express');
-
-var router = express.Router();
-var UserController = require('./userController');
-
-// Authorize each API with middleware and map to the Controller Functions
-router.post('/registration', UserController.createUser);
-router.post('/login/', UserController.loginUser);
-// on ne fait pas de suppression, et on n'a pas besoin d'accéder à la liste utilisateurs
-
-// Export the Router
-
-module.exports = router;
